Fix the deploy test
There was an attempt at a mock which isn't required. We can just
use the dqlite suite in this location.
